requirements.txt
Django==2.2.5
django-filter==2.3.0
djangorestframework==3.11.1
pytz==2020.1
sqlparse==0.3.1
